+++ shaders/tree_f.glsl
@@ -18,7 +18,7 @@ varying float isBark;
 
 varying vec4 pointLight [10];
 uniform vec4 light_color[10];
-uniform float num_lights;
+uniform int num_lights;
 
 void main()
     {
